refactor the commands
