+++ .github/workflows/test-fsat.yaml
@@ -31,6 +31,19 @@ jobs:
       - run: just test-appdev
         working-directory: full-stack-asset-transfer-guide
 
+  appdev-go:
+    runs-on: ${{ github.repository == 'hyperledger/fabric-samples' && 'fabric-ubuntu-22.04' || 'ubuntu-22.04' }}
+    steps:
+      - uses: actions/checkout@v6
+      - name: Set up Go
+        uses: actions/setup-go@v6
+        with:
+          go-version-file: full-stack-asset-transfer-guide/applications/trader-go/go.mod
+      - name: Set up Full Stack Runtime
+        uses: ./.github/actions/fsat-setup
+      - run: just test-appdev-go
+        working-directory: full-stack-asset-transfer-guide
+
   chaincode:
     runs-on: ${{ github.repository == 'hyperledger/fabric-samples' && 'fabric-ubuntu-22.04' || 'ubuntu-22.04' }}
     steps:
