+++ full-stack-asset-transfer-guide/applications/trader-go/commands/commands.go
@@ -1,10 +1,10 @@
-package main
+package commands
 
 import "github.com/hyperledger/fabric-gateway/pkg/client"
 
 type Command func(gw *client.Gateway, args []string) error
 
-var commands = map[string]Command{
+var Commands = map[string]Command{
 	"create":       cmdCreate,
 	"delete":       cmdDelete,
 	"getAllAssets": cmdGetAllAssets,
