+++ full-stack-asset-transfer-guide/applications/trader-go/commands/contract.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import (
 	"encoding/json"
