+++ full-stack-asset-transfer-guide/applications/trader-go/commands/expectedError.go
@@ -1,10 +1,4 @@
-/*
- * Copyright IBM Corp. All Rights Reserved.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package main
+package commands
 
 // ExpectedError represents a known, expected application error that should be
 // displayed normally rather than treated as an unexpected failure.
